+++ exporter-support/src/main/java/io/opentelemetry/auto/exportersupport/ConfigProvider.java
@@ -5,7 +5,7 @@ public interface ConfigProvider {
 
   int getInt(String key, int defaultValue);
 
-  long getLong(String key, int defaultValue);
+  long getLong(String key, long defaultValue);
 
   boolean getBoolean(String key, boolean defaultValue);
